Refactor to simplify lib use
* Migrate to `thiserror`
* Rename `Error` to `SdfGlyphError`
* Re-export freetype crate to simplify reuse and avoid compat issues
* Converted all asserts into proper results, libs should not panic
* Remove `freetype-rs` as a visible feature (use `freetype` as before)
* Bump version to 0.6 to indicate breaking change
* Refactor CI to streamline all testing/clippy/fmt/publishing
* Move code out of lib.rs into core/error/ft
* Move test data into `fixtures/` dir
* Move test into core.rs to keep tests together with relevant code
* rename `SDFGlyph -> SdfGlyph` for ease-of-reading
* Made `render_sdf` into a member function of bitmap
