+++ .github/workflows/ci.yml
@@ -1,0 +1,42 @@
+name: CI
+
+on:
+  push:
+    branches: [main]
+  pull_request:
+    branches: [main]
+  release:
+    types: [published]
+  workflow_dispatch:
+
+jobs:
+  build_and_test:
+    name: Run CI
+    runs-on: ubuntu-latest
+    steps:
+      - name: Checkout
+        uses: actions/checkout@v3
+      - name: Report Rust version
+        run: |
+          rustc --version
+          cargo --version
+      - name: Run fmt
+        run: |
+          cargo fmt --all -- --check
+      - name: Build & test
+        run: |
+          cargo build --release --all-features
+          cargo test --release --all-features
+      - name: Test docs
+        run: |
+          cargo test --release --doc
+          RUSTDOCFLAGS="-D warnings" cargo doc --release --no-deps
+      - name: Test clippy
+        run: |
+          cargo clippy --release --all-features -- -D warnings
+      - name: Publish on Release
+        if: github.event_name == 'release'
+        env:
+          CARGO_REGISTRY_TOKEN: ${{ secrets.CARGO_REGISTRY_TOKEN }}
+        run: |
+          cargo publish --all-features
