--- .github/workflows/publish.yml
@@ -1,24 +1,0 @@
-on:
-  release:
-    types: [published]
-
-name: Cargo Publish
-
-jobs:
-  publish:
-    name: Publish
-    runs-on: ubuntu-latest
-    steps:
-      - name: Checkout Repo
-        uses: actions/checkout@v2
-      - name: Use Rust Stable
-        uses: actions-rs/toolchain@v1
-        with:
-          toolchain: stable
-      - name: Publish
-        uses: actions-rs/cargo@v1
-        env:
-          CARGO_REGISTRY_TOKEN: ${{ secrets.CARGO_REGISTRY_TOKEN }}
-        with:
-          command: publish
-          args: --all-features
