--- .github/workflows/test.yml
@@ -1,17 +1,0 @@
-on: [push]
-
-name: Cargo Test
-
-jobs:
-  build_and_test:
-    name: Run unit tests
-    runs-on: ubuntu-latest
-    steps:
-      - uses: actions/checkout@v2
-      - uses: actions-rs/toolchain@v1
-        with:
-          toolchain: stable
-      - uses: actions-rs/cargo@v1
-        with:
-          command: test
-          args: --release --all-features
