+++ fixtures/glyph_alpha.json
@@ -1,0 +1,552 @@
+[
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  27,
+  75,
+  91,
+  55,
+  2,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  3,
+  141,
+  246,
+  196,
+  180,
+  231,
+  205,
+  38,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  117,
+  236,
+  42,
+  0,
+  0,
+  10,
+  180,
+  207,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  203,
+  130,
+  0,
+  0,
+  0,
+  0,
+  46,
+  255,
+  29,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  215,
+  107,
+  0,
+  0,
+  0,
+  0,
+  33,
+  255,
+  31,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  163,
+  169,
+  0,
+  0,
+  0,
+  0,
+  113,
+  221,
+  1,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  43,
+  247,
+  71,
+  0,
+  0,
+  65,
+  240,
+  71,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  96,
+  241,
+  69,
+  138,
+  231,
+  71,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  8,
+  199,
+  255,
+  191,
+  22,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  60,
+  222,
+  186,
+  136,
+  242,
+  58,
+  0,
+  0,
+  0,
+  0,
+  0,
+  8,
+  32,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  85,
+  240,
+  92,
+  0,
+  0,
+  117,
+  241,
+  56,
+  0,
+  0,
+  0,
+  0,
+  105,
+  175,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  17,
+  240,
+  79,
+  0,
+  0,
+  0,
+  0,
+  118,
+  240,
+  54,
+  0,
+  0,
+  0,
+  199,
+  81,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  100,
+  210,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  118,
+  239,
+  52,
+  0,
+  73,
+  230,
+  6,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  134,
+  173,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  119,
+  238,
+  64,
+  226,
+  84,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  123,
+  196,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  120,
+  255,
+  182,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  54,
+  252,
+  45,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  7,
+  163,
+  230,
+  235,
+  46,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  159,
+  230,
+  92,
+  4,
+  0,
+  0,
+  14,
+  90,
+  218,
+  189,
+  17,
+  120,
+  235,
+  44,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  3,
+  118,
+  231,
+  251,
+  222,
+  227,
+  251,
+  197,
+  91,
+  2,
+  0,
+  0,
+  125,
+  233,
+  43,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  1,
+  34,
+  55,
+  49,
+  17,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0,
+  0
+]
