+++ fixtures/glyph_sdf_u8.json
@@ -1,0 +1,552 @@
+[
+  0,
+  10,
+  31,
+  47,
+  58,
+  75,
+  89,
+  94,
+  95,
+  95,
+  95,
+  94,
+  89,
+  75,
+  55,
+  31,
+  10,
+  0,
+  0,
+  0,
+  0,
+  0,
+  4,
+  31,
+  55,
+  75,
+  89,
+  100,
+  118,
+  126,
+  127,
+  127,
+  126,
+  125,
+  118,
+  99,
+  75,
+  55,
+  31,
+  5,
+  0,
+  0,
+  0,
+  0,
+  18,
+  47,
+  75,
+  99,
+  118,
+  125,
+  144,
+  156,
+  158,
+  159,
+  158,
+  155,
+  143,
+  119,
+  100,
+  75,
+  48,
+  19,
+  0,
+  0,
+  0,
+  0,
+  27,
+  58,
+  89,
+  118,
+  143,
+  155,
+  159,
+  178,
+  184,
+  186,
+  182,
+  175,
+  157,
+  144,
+  119,
+  89,
+  59,
+  31,
+  4,
+  0,
+  0,
+  0,
+  31,
+  62,
+  94,
+  125,
+  155,
+  175,
+  192,
+  206,
+  199,
+  197,
+  204,
+  200,
+  180,
+  157,
+  126,
+  100,
+  75,
+  48,
+  19,
+  0,
+  0,
+  0,
+  31,
+  63,
+  95,
+  127,
+  159,
+  189,
+  204,
+  180,
+  159,
+  159,
+  176,
+  197,
+  201,
+  159,
+  144,
+  118,
+  89,
+  59,
+  28,
+  0,
+  0,
+  0,
+  31,
+  63,
+  95,
+  127,
+  159,
+  200,
+  191,
+  159,
+  144,
+  143,
+  157,
+  181,
+  223,
+  178,
+  157,
+  126,
+  94,
+  63,
+  31,
+  0,
+  0,
+  0,
+  31,
+  63,
+  95,
+  127,
+  159,
+  202,
+  188,
+  159,
+  127,
+  126,
+  157,
+  179,
+  223,
+  179,
+  157,
+  126,
+  94,
+  63,
+  31,
+  19,
+  5,
+  0,
+  31,
+  63,
+  95,
+  127,
+  159,
+  195,
+  196,
+  159,
+  145,
+  145,
+  159,
+  189,
+  202,
+  175,
+  155,
+  125,
+  94,
+  63,
+  59,
+  48,
+  31,
+  10,
+  31,
+  63,
+  95,
+  126,
+  157,
+  180,
+  206,
+  184,
+  158,
+  159,
+  183,
+  205,
+  184,
+  158,
+  143,
+  118,
+  94,
+  94,
+  89,
+  75,
+  55,
+  31,
+  31,
+  59,
+  89,
+  119,
+  144,
+  159,
+  187,
+  205,
+  183,
+  192,
+  204,
+  184,
+  158,
+  145,
+  119,
+  118,
+  125,
+  126,
+  118,
+  100,
+  75,
+  48,
+  55,
+  76,
+  100,
+  119,
+  145,
+  158,
+  176,
+  200,
+  223,
+  199,
+  178,
+  158,
+  145,
+  119,
+  118,
+  143,
+  156,
+  157,
+  144,
+  118,
+  89,
+  59,
+  75,
+  100,
+  119,
+  145,
+  158,
+  182,
+  203,
+  198,
+  192,
+  205,
+  182,
+  158,
+  145,
+  119,
+  125,
+  156,
+  176,
+  179,
+  157,
+  126,
+  94,
+  63,
+  89,
+  118,
+  144,
+  158,
+  185,
+  205,
+  186,
+  159,
+  159,
+  189,
+  205,
+  182,
+  158,
+  145,
+  127,
+  159,
+  188,
+  197,
+  159,
+  127,
+  95,
+  63,
+  94,
+  126,
+  156,
+  177,
+  205,
+  185,
+  159,
+  145,
+  146,
+  159,
+  190,
+  205,
+  182,
+  158,
+  145,
+  159,
+  200,
+  185,
+  158,
+  127,
+  95,
+  63,
+  95,
+  127,
+  159,
+  187,
+  201,
+  159,
+  145,
+  119,
+  119,
+  146,
+  159,
+  190,
+  205,
+  181,
+  158,
+  184,
+  204,
+  176,
+  155,
+  125,
+  94,
+  62,
+  95,
+  127,
+  159,
+  192,
+  196,
+  159,
+  127,
+  100,
+  101,
+  119,
+  146,
+  159,
+  190,
+  205,
+  183,
+  203,
+  185,
+  158,
+  143,
+  118,
+  89,
+  58,
+  95,
+  127,
+  159,
+  190,
+  199,
+  159,
+  145,
+  125,
+  118,
+  118,
+  125,
+  146,
+  159,
+  190,
+  223,
+  198,
+  159,
+  145,
+  119,
+  100,
+  75,
+  55,
+  95,
+  126,
+  158,
+  182,
+  206,
+  180,
+  159,
+  155,
+  143,
+  143,
+  156,
+  159,
+  176,
+  195,
+  204,
+  204,
+  181,
+  157,
+  144,
+  119,
+  100,
+  75,
+  90,
+  119,
+  145,
+  159,
+  195,
+  204,
+  186,
+  175,
+  159,
+  159,
+  177,
+  186,
+  202,
+  198,
+  177,
+  190,
+  204,
+  180,
+  157,
+  144,
+  119,
+  89,
+  75,
+  100,
+  125,
+  155,
+  175,
+  190,
+  204,
+  206,
+  203,
+  203,
+  206,
+  199,
+  186,
+  175,
+  156,
+  159,
+  190,
+  204,
+  180,
+  157,
+  126,
+  95,
+  58,
+  89,
+  118,
+  143,
+  155,
+  159,
+  175,
+  179,
+  182,
+  181,
+  177,
+  159,
+  159,
+  155,
+  143,
+  146,
+  159,
+  159,
+  157,
+  144,
+  119,
+  89,
+  47,
+  75,
+  99,
+  118,
+  125,
+  143,
+  155,
+  157,
+  158,
+  157,
+  156,
+  144,
+  127,
+  125,
+  118,
+  119,
+  127,
+  127,
+  126,
+  119,
+  100,
+  75,
+  31,
+  55,
+  75,
+  89,
+  99,
+  118,
+  125,
+  126,
+  126,
+  126,
+  126,
+  118,
+  100,
+  94,
+  89,
+  90,
+  95,
+  95,
+  95,
+  89,
+  75,
+  55,
+  10,
+  31,
+  47,
+  58,
+  75,
+  89,
+  94,
+  94,
+  95,
+  95,
+  94,
+  89,
+  75,
+  62,
+  58,
+  59,
+  63,
+  63,
+  63,
+  59,
+  48,
+  31
+]
